feature/SA-39 HeaderMenu component created
